Added the UnitController.cs to allow the player to select an unit in the game, but can be altered later to instead allow for the player to only be able to move their units.
I added an interface named IGetInput.cs which provides a ray and raycasthit return method type which means so as a long as any class uses those functions such as the KeyboardInput.cs it can be used as input dynamically.

The colour changing of the tiles now also works and resets correctly since the colour of the tile is now stored in the Tile.cs class.