Made the controller have an icon on the UI which is moved and then translated from screen space to world space which gives the desired controls that I wanted beforehand.
Moved some functions from KeyboardInput.cs to UserInput.cs since I found I also needed them as well as the camera data type in the ControllerInput.cs.

Changed the events in the UnitController.cs to set the new event rather than add another to the event list to be called since it was causing the issue where if you would click on the same unit twice in a row it would just keep adding the units as listeners at the same time.

Changed the events where appropriate from (if <event> != null) to instead use the syntax of <event>?Invoke() which is just better shorthand in my opinion.