Updated GDD - Actually explains the game now.
Updated the GDD to explain what the game is and what it is about, I hope people are able to understand what the Designers were roughly planning. Please let us know if you have any questions, ideas or dislikes to what is currently planned. Also tell us if you still don't really understand it, I'm happy to go through it with you.